requirements.txt
anthropic>=0.25.0
fastapi==0.115.6
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
uvicorn[standard]==0.34.0
